% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ae_tables.R
\name{create_ae_tables}
\alias{create_ae_tables}
\title{Create AE tables}
\usage{
create_ae_tables(ae_data)
}
\arguments{
\item{ae_data}{A string containing the path to AE data in csv format, or the
AE data as a data frame.}
}
\value{
A list containing the four AE reporting tables.
}
\description{
DESCRIPTION
}
